feat(parkit): add 'parkit ls' subcommand and extract shared ls logic
Idea 4 (--json output) + Idea 5 (shared backend) from the parkit ls
brainstorm.

Changes
-------
src/parkit/src/utils.py
  - Add human_size() — human-readable byte size string
  - Add collection_exists() — dm_get_collection existence gate

src/parkit/src/lscollection.py  (new)
  - _LS_CRITERIA, parse_deposit_date(), query_all_dataobjects(),
    render_ls_tree(), render_ls_json(), ls_collection()
  - ls_collection(collection_path, json_output=False) is the single
    public entry point used by both parkit ls and projark ls

src/parkit/projark.py
  - Import human_size, collection_exists from utils;
    import ls_collection, query_all_dataobjects from lscollection
  - _human_size() and _collection_exists() become thin aliases
  - _run_retrieve() now calls query_all_dataobjects() from shared module
  - _run_ls() simplified to ~10 lines; all rendering delegated to
    ls_collection()
  - Remove ~120 lines of duplicated ls logic
  - Add --json flag to 'projark ls' subparser

src/parkit/__main__.py
  - Import ls_collection, collection_exists, sys
  - Add 'ls' subparser: positional COLLECTION_PATH + --json flag
  - Add 'ls' to subcommands list and version loop
  - Dispatch: existence gate + ls_collection()

Usage
-----
  parkit ls /CCBR_Archive/GRIDFTP/Project_CCBR-982
  parkit ls /CCBR_Archive/GRIDFTP/Project_CCBR-982 --json
  projark ls 982
  projark ls 982 --json

Closes #62

⚡ Generated using AI ⚡
